requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
edge-tts>=1.0.0
chromadb>=0.4.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
python-pptx>=0.6.21
